+++ packages/data-model/index.ts
@@ -1,2 +1,3 @@
+export * from './src/config/db.js';
 export * from './src/functions/migrate.js';
-export * from './src/models/index.js';
+export * as models from './src/models/index.js';
