+++ packages/data-model/src/functions/migrate.ts
@@ -1,28 +1,20 @@
-import 'dotenv/config';
-
 import { drizzle } from 'drizzle-orm/node-postgres';
 import { migrate as migrator } from 'drizzle-orm/node-postgres/migrator';
 import path from 'path';
 import pg from 'pg';
 import { fileURLToPath } from 'url';
 
-const getRequiredConfig = (name: string) => {
-	const value = process.env[name];
-	if (!value) {
-		throw new Error(`No Environment Variable provided for required configuration parameter '${name}'`);
-	}
-	return value;
-};
+import type { DbConfig } from '../config/db.js';
 
-export async function migrate() {
+export async function migrate(config: DbConfig) {
 	console.log('Running your migrations...');
 
 	const sql = new pg.Pool({
-		host: getRequiredConfig('DB_HOST'),
-		database: getRequiredConfig('DB_NAME'),
-		password: getRequiredConfig('DB_PASSWORD'),
-		port: Number(getRequiredConfig('DB_PORT')),
-		user: getRequiredConfig('DB_USER'),
+		host: config.host,
+		database: config.database,
+		password: config.password,
+		port: config.port,
+		user: config.user,
 	});
 	const db = drizzle(sql);
